fix(lint): clear ruff-blocking findings across src/mlframe (60 -> 0)
F401 (29): removed genuinely-dead typing/module imports. Two files had
load-bearing re-exports mixed in with real dead imports -- verified via
grep for external `from module import name` usages before touching
anything: models/optimization.py re-exports 5 names from
._optimization_shared "for BC" per its own comment, and
feature_selection/filters/batch_pair_mi_gpu.py re-exports 5 names from
._batch_pair_mi_cuda_kernels consumed directly by 13 test/benchmark
files. Both get noqa: F401 (matching the existing _SearchNotReady/
_CUDA_AVAIL precedent in the same files) instead of deletion; only the
3 genuinely-unused imports in the GPU file (math, typing.Any, numba)
were removed.

RUF005/RUF046/RUF059/B007/PERF401/PERF403 (31, via safe + reviewed
unsafe --fix, plus 4 hand-fixes for unpacked-var/loop-var renames and
list.extend/dict-comprehension rewrites): behavior-preserving perf/
style cleanups.

RUF013 (1): as_of_aggregate.py's query_entity_col got an explicit
Optional[str] (was implicit-Optional with a now-stale type-ignore that
mypy flagged as unused once ruff added the annotation -- removed).

PERF203 (1): polynom_pair_fe.py's per-column dtype-coercion try/except
inside a loop is intentional (runs once per FE round, not a hot path);
noqa'd with the existing rationale from its own surrounding comments.

Also closed the resulting interrogate docstring-coverage gaps (100%
gate, whole-file) surfaced by touching these 15 files: one-line
docstrings on the previously-undocumented functions/methods/classes in
each, no behavior changes.

Verified: ruff/mypy/black clean on every touched file; full biz_val +
unit test coverage for every touched module passes (run in batches to
avoid resource contention with concurrent training jobs elsewhere in
this shared working tree).

+++ src/mlframe/feature_selection/drop_near_noise_univariate_auc.py
@@ -78,7 +78,7 @@ def drop_near_noise_univariate_auc(
         raise ValueError(f"n_bootstrap must be a positive int when provided, got {n_bootstrap}")
 
     n_rows = X.shape[0]
-    sample_size = max(2, int(round(n_rows * bootstrap_frac)))
+    sample_size = max(2, round(n_rows * bootstrap_frac))
     rng = np.random.default_rng(random_state)
 
     resampled_aucs = np.empty((n_bootstrap, len(cols)), dtype=np.float64)
